requirements.txt
requests
git+https://github.com/Hrishi2861/pyrofork-2.2.11-peer-fix.git
python-dotenv
TgCrypto
bs4
vcsi
playwright
myjdapi